// src/pages/Contact/type.ts
export interface ContactLanguage {
  p1first: Translate;
  p1end: Translate;
  p2star: Translate;
  p22: Translate;
  p23: Translate;
  p24: Translate;
  p25: Translate;
  p26: Translate;
  p27: Translate;
  p31: Translate;
  p32: Translate;
  p33: Translate;
  p41: Translate;
  p42: Translate;
  p43: Translate;
}

export interface Translate {
  fr: string;
  en: string;
}